backend/requirements.txt
gunicorn
Flask==3.0.0
Flask-Cors==4.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-cloud-aiplatform==1.78.0
google-cloud-storage==2.19.0
google-generativeai==0.8.4
google-cloud-bigquery
db_dtypes==1.4.0
numpy==1.26.0
pandas==2.1.0
plotly==5.24.1
requests==2.31.0
PyMySQL==1.1.1
python-dotenv==1.0.0
tiktoken==0.8.0
